requirements.txt
flask
uuid
redis
pyjwt
flask-cors
pytube

